+++ R/Set_Interval_SpecialSet.R
@@ -273,37 +273,37 @@ ExtendedReals <- R6Class("ExtendedReals",inherit = Reals,
 #' @family special sets
 #' @export
 Complex <- R6Class("Complex",inherit = SpecialSet,
-                   public = list(
-                     #' @description Create a new `Complex` object.
-                     #' @return A new `Complex` object.
-                     #' @param lower complex. Where to start the set.
-                     #' @param upper complex. Where to end the set.
-                     initialize = function(lower = -Inf+0i, upper = Inf+0i){
-                       super$initialize(lower = as.complex(lower), upper = as.complex(upper), type = "()", class = "numeric")
-                     },
+  public = list(
+    #' @description Create a new `Complex` object.
+    #' @return A new `Complex` object.
+    #' @param lower complex. Where to start the set.
+    #' @param upper complex. Where to end the set.
+    initialize = function(lower = -Inf+0i, upper = Inf+0i){
+      super$initialize(lower = as.complex(lower), upper = as.complex(upper), type = "()", class = "numeric")
+    },
 
-                     #' @description Tests to see if \code{x} is contained in the Set.
-                     #'
-                     #' @param x any. Object or vector of objects to test.
-                     #' @param all logical. If `FALSE` tests each `x` separately. Otherwise returns `TRUE` only if all `x` pass test.
-                     #' @param bound logical.
-                     #'
-                     #' @details \code{x} can be of any type, including a Set itself. \code{x} should be a tuple if
-                     #' checking to see if it lies within a set of dimension greater than one. To test for multiple \code{x}
-                     #' at the same time, then provide these as a list.
-                     #'
-                     #' If `all = TRUE` then returns `TRUE` if all `x` are contained in the `Set`, otherwise
-                     #' returns a vector of logicals. For [Interval]s, `bound` is used to specify if elements lying on the
-                     #' (possibly open) boundary of the interval are considered contained (`bound = TRUE`) or not (`bound = FALSE`).
-                     #'
-                     #' @return If \code{all} is `TRUE` then returns `TRUE` if all elements of \code{x} are contained in the `Set`, otherwise
-                     #' `FALSE.` If \code{all} is `FALSE` then returns a vector of logicals corresponding to each individual
-                     #' element of \code{x}.
-                     #'
-                     #' The infix operator `%inset%` is available to test if `x` is an element in the `Set`,
-                     #' see examples.
-                     contains = function(x, all = FALSE, bound = NULL){
-                       ret <- sapply(x, function(y) inherits(y, "complex"))
-                       returner(ret, all)
-                     }
-                   ))
+    #' @description Tests to see if \code{x} is contained in the Set.
+    #'
+    #' @param x any. Object or vector of objects to test.
+    #' @param all logical. If `FALSE` tests each `x` separately. Otherwise returns `TRUE` only if all `x` pass test.
+    #' @param bound logical.
+    #'
+    #' @details \code{x} can be of any type, including a Set itself. \code{x} should be a tuple if
+    #' checking to see if it lies within a set of dimension greater than one. To test for multiple \code{x}
+    #' at the same time, then provide these as a list.
+    #'
+    #' If `all = TRUE` then returns `TRUE` if all `x` are contained in the `Set`, otherwise
+    #' returns a vector of logicals. For [Interval]s, `bound` is used to specify if elements lying on the
+    #' (possibly open) boundary of the interval are considered contained (`bound = TRUE`) or not (`bound = FALSE`).
+    #'
+    #' @return If \code{all} is `TRUE` then returns `TRUE` if all elements of \code{x} are contained in the `Set`, otherwise
+    #' `FALSE.` If \code{all} is `FALSE` then returns a vector of logicals corresponding to each individual
+    #' element of \code{x}.
+    #'
+    #' The infix operator `%inset%` is available to test if `x` is an element in the `Set`,
+    #' see examples.
+    contains = function(x, all = FALSE, bound = NULL){
+      returner(x = sapply(x, is.complex),
+               all = all)
+    }
+  ))