+++ NEWS.md
@@ -1,5 +1,7 @@
 # set6 0.1.3.9000
 
+* Speed performance improvements for `$contains` method for `Interval` and `Set`. `Rcpp` now used for `Interval`.
+
 # set6 0.1.3
 
 * Added assertion for testing if a set is countably finite